requirements.txt
fastapi
uvicorn
pydantic
sentence_transformers
python-dotenv
streamlit
requests